pyproject.toml
[tool.poetry]
name = "lesson-27-homework"
version = "0.1.0"
description = ""
authors = ["Zidolch <yury.batin@gmail.com>"]
readme = "README.md"
packages = [{include = "lesson_27_homework"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
